# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FamilyBranch API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /family-units/{unitId}/tree:
    get:
      operationId: getFamilyTree
      tags: [family-groups]
      summary: Get the full family tree for a unit (unit + all linked units)
      parameters:
        - $ref: "#/components/parameters/unitId"
      responses:
        "200":
          description: Family tree data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FamilyTree"

  # ─── Summary / Dashboard ─────────────────────────────────────────
  /family-units/{unitId}/summary:
    get:
      operationId: getUnitSummary
      tags: [summary]
      summary: Get dashboard summary stats for a unit
      parameters:
        - $ref: "#/components/parameters/unitId"
      responses:
        "200":
          description: Summary stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnitSummary"
  /family-units/{unitId}/birthdays:
    get:
      operationId: getUpcomingBirthdays
      tags: [summary]
      summary: Get upcoming birthdays in the unit and linked units
      parameters:
        - $ref: "#/components/parameters/unitId"
      responses:
        "200":
          description: Upcoming birthdays
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BirthdayEntry"
components:
  parameters:
...
    unitId:
      name: unitId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
...
    Person:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        photoUrl:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        addressLine1:
          type: string
          nullable: true
        addressCity:
          type: string
          nullable: true
        addressState:
          type: string
          nullable: true
        addressZip:
          type: string
          nullable: true
        addressCountry:
          type: string
          nullable: true
        birthday:
          type: string
          nullable: true
        showBirthYear:
          type: boolean
        instagram:
          type: string
          nullable: true
        facebook:
          type: string
          nullable: true
        tiktok:
          type: string
          nullable: true
        linkedin:
          type: string
          nullable: true
        otherSocial:
          type: string
          nullable: true
        relationshipLabel:
          type: string
        familyUnitId:
          type: string
          format: uuid
        isAdmin:
          type: boolean
        claimed:
          type: boolean
        claimedAt:
          type: string
          format: date-time
          nullable: true
        inviteExpiresAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - firstName
        - lastName
        - relationshipLabel
        - familyUnitId
        - isAdmin
        - claimed
        - showBirthYear
        - createdAt
        - updatedAt
...
    FamilyTreeNode:
      type: object
      properties:
        unitId:
          type: string
          format: uuid
        unitName:
          type: string
        parentUnitId:
          type: string
          format: uuid
          nullable: true
        members:
          type: array
          items:
            $ref: "#/components/schemas/Person"
        children:
          type: array
          items:
            $ref: "#/components/schemas/FamilyTreeNode"
      required: [unitId, unitName, members, children]
    FamilyTree:
      type: object
      properties:
        rootUnit:
          $ref: "#/components/schemas/FamilyTreeNode"
        totalUnits:
          type: integer
        totalMembers:
          type: integer
      required: [rootUnit, totalUnits, totalMembers]

    # ─── Summary ───────────────────────────────────────────────────
    UnitSummary:
      type: object
      properties:
        unitId:
          type: string
          format: uuid
        unitName:
          type: string
        totalMembers:
          type: integer
        claimedMembers:
          type: integer
        unclaimedMembers:
          type: integer
        claimRate:
          type: number
          description: Percentage of members who have claimed their profile (0–100)
        linkedUnits:
          type: integer
        pendingInvites:
          type: integer
        pendingLinkRequests:
          type: integer
      required:
        - unitId
        - unitName
        - totalMembers
        - claimedMembers
        - unclaimedMembers
        - claimRate
        - linkedUnits
        - pendingInvites
        - pendingLinkRequests
    BirthdayEntry:
      type: object
      properties:
        personId:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        relationshipLabel:
          type: string
        birthday:
          type: string
        unitName:
          type: string
        daysUntil:
          type: integer
          description: Days until their next birthday
      required:
        - personId
        - firstName
        - lastName
        - relationshipLabel
        - birthday
        - unitName
        - daysUntil